+++ js/lib/NodeLoader.js
@@ -47,7 +47,7 @@ module.exports = function (ilib) {
         //console.log("base is defined as " + this.base);
 
         // this.includePath.push(path.join(this.root, "resources"));     // always check the application's resources dir first
-        this._exists(this.root, "resources"); // always check the application's resources dir first
+        this._exists(path.join(this.root, "resources")); // always check the application's resources dir first
         this._exists(path.join(this.root, "locale"), "localeinfo.json");
 
         // then a standard locale dir of a built version of ilib from npm