+++ docs/ReleaseNotes.md
@@ -10,6 +10,7 @@ New Features:
 Bug Fixes:
 * Updated to IANA time zone data 2021c
 * Updated to IANA time zone data 2021d
+* Updated to IANA time zone data 2021e
 
 Build 017
 -------